removed dsnname customization
