+++ config/crd/bases/persistance.atlas.infoblox.com_databaseclaims.yaml
@@ -91,7 +91,10 @@ spec:
                 type: string
               dsnName:
                 default: dsn.txt
-                description: DSN key name.
+                description: |-
+                  DSN is used to name the secret key that contains old style DSN for postgres.
+                  This field is deprecated, update your code to access key "dsn.txt" or
+                  preferrably "uri_dsn.txt" as this field is not customizable any longer.
                 type: string
               enableReplicationRole:
                 default: false
