+++ test/e2e/dbc_end2end_test.go
@@ -116,7 +116,6 @@ var _ = Describe("AWS", Ordered, func() {
 					SecretName:            "newdb-secret-db1",
 					Username:              "sample_user",
 					Type:                  "postgres",
-					DSNName:               "dsn",
 					EnableReplicationRole: ptr.To(false),
 					UseExistingSource:     ptr.To(false),
 				},
@@ -377,7 +376,6 @@ var _ = Describe("AWS", Ordered, func() {
 					SecretName:            "sample-secret-db2",
 					Username:              "sample_user",
 					Type:                  "postgres",
-					DSNName:               "dsn",
 					EnableReplicationRole: ptr.To(false),
 					UseExistingSource:     ptr.To(true),
 					DBVersion:             "15.5",
